refactor: move watchlistContext files into a watchlist folder
